DB-2149 support parsing epoch sec/millisec as timestamp
