+++ src/main/java/io/thekraken/grok/api/Converter.java
@@ -102,81 +102,4 @@ public class Converter {
   }
 }
 
-//
-// Converters
-//
-interface IConverter<T> {
-  T convert(CharSequence value);
 
-  default IConverter<T> newConverter(String param, Object... params) {
-    return this;
-  }
-}
-
-
-class DateConverter implements IConverter<Instant> {
-  private final DateTimeFormatter formatter;
-  private final ZoneId timeZone;
-  private final boolean useCache;
-  private final String pattern;
-  private final boolean containsYear;
-
-  private final Cache<CharSequence, Instant> timestampCache =
-      CacheBuilder.newBuilder().maximumSize(1000).build();
-
-  public DateConverter() {
-    this.formatter = DateTimeFormatter.ISO_DATE_TIME;
-    this.timeZone = ZoneOffset.UTC;
-    this.useCache = true;
-    this.pattern = null;
-    this.containsYear = true;
-  }
-
-  private DateConverter(String pattern, ZoneId timeZone) {
-    this.formatter = DateTimeFormatter.ofPattern(pattern);
-    this.timeZone = timeZone;
-    this.useCache = !(pattern.contains("S") || pattern.contains("n") || pattern.contains("N") || pattern.contains("A"));
-    this.pattern = pattern;
-    var format = pattern.toLowerCase();
-    this.containsYear = format.contains("u") || format.contains("y");
-  }
-
-  @Override
-  public Instant convert(CharSequence value) {
-    if (useCache) {
-      var ts = timestampCache.getIfPresent(value);
-      if (ts != null) {
-        return ts;
-      }
-    }
-
-    var formatter = this.formatter;
-
-    if (!containsYear) {
-      LocalDate today = LocalDate.now(timeZone);
-      formatter = new DateTimeFormatterBuilder().append(this.formatter)
-              .parseDefaulting(ChronoField.YEAR, today.getYear())
-              .parseDefaulting(ChronoField.MONTH_OF_YEAR, today.getMonthValue())
-              .parseDefaulting(ChronoField.DAY_OF_MONTH, today.getDayOfMonth())
-              .toFormatter().withZone(timeZone);
-    }
-
-    TemporalAccessor dt = formatter.parseBest(value, ZonedDateTime::from, LocalDateTime::from);
-    Instant ts;
-    if (dt instanceof ZonedDateTime) {
-      ts =  ((ZonedDateTime)dt).toInstant();
-    } else {
-      ts = ((LocalDateTime) dt).atZone(timeZone).toInstant();
-    }
-    if (useCache) {
-      timestampCache.put(value, ts);
-    }
-    return ts;
-  }
-
-  @Override
-  public DateConverter newConverter(String param, Object... params) {
-    Preconditions.checkArgument(params.length == 1 && params[0] instanceof ZoneId);
-    return new DateConverter(param, (ZoneId) params[0]);
-  }
-}
