+++ src/main/java/io/thekraken/grok/api/IConverter.java
@@ -1,0 +1,9 @@
+package io.thekraken.grok.api;
+
+public interface IConverter<T> {
+  T convert(CharSequence value);
+
+  default IConverter<T> newConverter(String param, Object... params) {
+    return this;
+  }
+}
